Refactor API routes for consistency
Reorganize API routes to more logically represent resources in a
hierarchy. Move community enpoints that operate on the community the
apikey is valid on to /communities/own, merge apikey/create and
apikey/revoke into one route, move discord related routes to /discord/,
consolidate all fetch multiple reports routes into /reports/, and
likewise with /revocations/.
