+++ src/routes/revocations.controller.ts
@@ -15,6 +15,7 @@ import {
 } from "fagc-api-types"
 import { DocumentType } from "@typegoose/typegoose"
 import { BeAnObject } from "@typegoose/typegoose/lib/types"
+import validator from "validator"
 
 @Controller({ route: "/revocations" })
 export default class RevocationController {
@@ -22,7 +23,20 @@ export default class RevocationController {
 		url: "/",
 		options: {
 			schema: {
-				description: "Fetch all revocations",
+				querystring: z.object({
+					playername: z.string().or(z.array(z.string())),
+					categoryId: z.string().or(z.array(z.string())),
+					adminId: z.string().or(z.array(z.string())),
+					after: z.string().optional().refine(
+						(input) => input === undefined || validator.isISO8601(input),
+						"Invalid timestamp"
+					),
+					revokedAfter: z.string().optional().refine(
+						(input) => input === undefined || validator.isISO8601(input),
+						"Invalid timestamp"
+					),
+				}),
+				description: "Fetch revocations from your community",
 				tags: [ "revocations" ],
 				security: [
 					{
@@ -42,9 +56,18 @@ export default class RevocationController {
 	})
 	@Authenticate
 	async fetchAll(
-		req: FastifyRequest,
+		req: FastifyRequest<{
+			Querystring: {
+				playername?: string | string[],
+				categoryId?: string | string[],
+				adminId?: string | string[],
+				after?: string,
+				revokedAfter?: string,
+			}
+		}>,
 		res: FastifyReply
 	): Promise<FastifyReply> {
+		const { playername, categoryId, adminId, after, revokedAfter } = req.query
 		const community = req.requestContext.get("community")
 		if (!community) return res.status(404).send({
 			errorCode: 404,
@@ -53,8 +76,15 @@ export default class RevocationController {
 		})
 
 		const revocations = await ReportInfoModel.find({
+			playername: playername,
+			categoryId: categoryId,
+			adminId: adminId,
+			...(after ? { createdAt: { $gt: new Date(after) } } : {}),
 			communityId: community.id,
-			revokedAt: { $ne: null },
+			revokedAt: {
+				$ne: null,
+				...(revokedAfter ? { $gt: new Date(revokedAfter) } : {}),
+			},
 		})
 
 		return res.send(revocations)
@@ -68,7 +98,7 @@ export default class RevocationController {
 					id: z.string(),
 				}),
 
-				description: "Fetch a revocation by ID",
+				description: "Fetch revocation",
 				tags: [ "revocations" ],
 				security: [
 					{
@@ -77,14 +107,14 @@ export default class RevocationController {
 				],
 				response: {
 					"200": {
-						allOf: [ { nullable: true }, { $ref: "ReportInfoClass#" } ],
+						allOf: [ { nullable: true }, { $ref: "ReportClass#" } ],
 					},
 				},
 			},
 		},
 	})
 	@Authenticate
-	async fetchId(
+	async fetch(
 		req: FastifyRequest<{
 			Params: {
 				id: string
@@ -111,13 +141,11 @@ export default class RevocationController {
 	}
 
 	@POST({
-		url: "/:id",
+		url: "/",
 		options: {
 			schema: {
-				params: z.object({
-					id: z.string(),
-				}),
 				body: z.object({
+					reportId: z.string(),
 					adminId: z.string(),
 				}),
 
@@ -139,10 +167,8 @@ export default class RevocationController {
 	@Authenticate
 	async revokeReport(
 		req: FastifyRequest<{
-			Params: {
-				id: string
-			}
 			Body: {
+				reportId: string
 				adminId: string
 			}
 		}>,
@@ -157,7 +183,7 @@ export default class RevocationController {
 			})
 
 		const report = await ReportInfoModel.findOne({
-			id: req.params.id,
+			id: req.body.reportId,
 			revokedAt: { $exists: false },
 		})
 		if (!report)
@@ -219,58 +245,6 @@ export default class RevocationController {
 		return res.status(200).send(revocation)
 	}
 
-	@GET({
-		url: "/category/:id",
-		options: {
-			schema: {
-				params: z.object({
-					id: z.string(),
-				}),
-
-				description: "Fetch all revocations of a category in your community",
-				tags: [ "revocations" ],
-				security: [
-					{
-						authorization: [],
-					},
-				],
-				response: {
-					"200": {
-						type: "array",
-						items: {
-							$ref: "RevocationClass#",
-						},
-					},
-				},
-			},
-		},
-	})
-	@Authenticate
-	async fetchCategory(
-		req: FastifyRequest<{
-			Params: {
-				id: string
-			}
-		}>,
-		res: FastifyReply
-	): Promise<FastifyReply> {
-		const { id } = req.params
-		const community = req.requestContext.get("community")
-		if (!community) return res.status(404).send({
-			errorCode: 404,
-			error: "Community not found",
-			message: "Community not found",
-		})
-
-		const revocations = await ReportInfoModel.find({
-			categoryIdId: id,
-			communityId: community.id,
-			revokedAt: { $ne: null },
-		})
-		
-		return res.send(revocations)
-	}
-
 	@POST({
 		url: "/category/:id",
 		options: {
@@ -301,7 +275,7 @@ export default class RevocationController {
 		},
 	})
 	@Authenticate
-	async revokeCategory(
+	async revokeByCategory(
 		req: FastifyRequest<{
 			Params: {
 				id: string
@@ -394,58 +368,6 @@ export default class RevocationController {
 		return res.status(200).send(revocations)
 	}
 
-	@GET({
-		url: "/player/:playername",
-		options: {
-			schema: {
-				params: z.object({
-					playername: z.string(),
-				}),
-
-				description: "Fetch all revocations of a player in your community",
-				tags: [ "revocations" ],
-				security: [
-					{
-						authorization: [],
-					},
-				],
-				response: {
-					"200": {
-						type: "array",
-						items: {
-							$ref: "RevocationClass#",
-						},
-					},
-				},
-			},
-		},
-	})
-	@Authenticate
-	async fetchPlayer(
-		req: FastifyRequest<{
-			Params: {
-				playername: string
-			}
-		}>,
-		res: FastifyReply
-	): Promise<FastifyReply> {
-		const { playername } = req.params
-		const community = req.requestContext.get("community")
-		if (!community) return res.status(404).send({
-			errorCode: 404,
-			error: "Community not found",
-			message: "Community not found",
-		})
-
-		const revocations = await ReportInfoModel.find({
-			playername: playername,
-			communityId: community.id,
-			revokedAt: { $ne: null },
-		})
-
-		return res.send(revocations)
-	}
-
 	@POST({
 		url: "/player/:playername",
 		options: {
@@ -476,7 +398,7 @@ export default class RevocationController {
 		},
 	})
 	@Authenticate
-	async revokePlayer(
+	async revokeByPlayer(
 		req: FastifyRequest<{
 			Params: {
 				playername: string
@@ -569,66 +491,6 @@ export default class RevocationController {
 		return res.status(200).send(revocations)
 	}
 
-	@GET({
-		url: "/admin/:snowflake",
-		options: {
-			schema: {
-				params: z.object({
-					snowflake: z.string(),
-				}),
-
-				description: "Fetch all revocations revoked by an admin in your community",
-				tags: [ "revocations" ],
-				security: [
-					{
-						authorization: [],
-					},
-				],
-				response: {
-					"200": {
-						type: "array",
-						items: {
-							$ref: "RevocationClass#",
-						},
-					},
-				},
-			},
-		},
-	})
-	@Authenticate
-	async fetchAdmin(
-		req: FastifyRequest<{
-			Params: {
-				snowflake: string
-			}
-		}>,
-		res: FastifyReply
-	): Promise<FastifyReply> {
-		const { snowflake } = req.params
-		const community = req.requestContext.get("community")
-		if (!community) return res.status(404).send({
-			errorCode: 404,
-			error: "Community not found",
-			message: "Community not found",
-		})
-
-		const admin = await validateDiscordUser(snowflake)
-		if (!admin)
-			return res.status(400).send({
-				errorCode: 400,
-				error: "Bad Request",
-				message: "adminId must be a valid Discord user ID",
-			})
-
-		const revocations = await ReportInfoModel.find({
-			adminId: snowflake,
-			communityId: community.id,
-			revokedAt: { $ne: null },
-		})
-		
-		return res.send(revocations)
-	}
-
 	@POST({
 		url: "/admin/:snowflake",
 		options: {
@@ -656,7 +518,7 @@ export default class RevocationController {
 		},
 	})
 	@Authenticate
-	async revokeAdmin(
+	async revokeByAdmin(
 		req: FastifyRequest<{
 			Params: {
 				snowflake: string
@@ -736,51 +598,6 @@ export default class RevocationController {
 			})
 		})
 		
-		return res.send(revocations)
-	}
-
-	@GET({
-		url: "/since/:timestamp",
-		options: {
-			schema: {
-				params: z.object({
-					timestamp: z.string(), // TODO: better time validation
-				}),
-
-				description:
-					"Fetch all revocations of a player modified since a timestamp",
-				tags: [ "revocations" ],
-				security: [
-					{
-						authorization: [],
-					},
-				],
-				response: {
-					"200": {
-						type: "array",
-						items: {
-							$ref: "RevocationClass#",
-						},
-					},
-				},
-			},
-		},
-	})
-	async getModifiedSince(
-		req: FastifyRequest<{
-			Params: {
-				timestamp: string
-			}
-		}>,
-		res: FastifyReply
-	): Promise<FastifyReply> {
-		const { timestamp } = req.params
-
-		const date = new Date(timestamp)
-
-		const revocations = await ReportInfoModel.find({
-			revokedAt: { $gt: date },
-		})
 		return res.send(revocations)
 	}
 }
